refactor: component hierarchy
